Creating default project for users
