+++ litmus-portal/frontend/src/views/GetStarted/PasswordSet/index.tsx
@@ -1,9 +1,24 @@
-/* eslint-disable react/no-danger */
+import { useLazyQuery, useMutation } from '@apollo/client';
 import { Typography } from '@material-ui/core';
 import { ButtonFilled, InputField } from 'litmus-ui';
-import React, { useRef } from 'react';
+import React, { useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { getUsername } from '../../../utils/auth';
+import Loader from '../../../components/Loader';
+import config from '../../../config';
+import { CREATE_PROJECT, CREATE_USER, GET_USER_INFO } from '../../../graphql';
+import {
+  CreateUserData,
+  CurrentUserDetails,
+  Project,
+} from '../../../models/graphql/user';
+import {
+  getToken,
+  getUserEmail,
+  getUserFullName,
+  getUserId,
+  getUsername,
+  getUserRole,
+} from '../../../utils/auth';
 import { validateConfirmPassword } from '../../../utils/validate';
 import useStyles from './styles';
 
@@ -12,18 +27,7 @@ interface PasswordReset {
   confirmPassword: string;
 }
 
-interface PasswordSetProps {
-  handleNext: () => void;
-  currentStep: number;
-  totalStep: number;
-  setPassword: (e: any) => void;
-}
-const PasswordSet: React.FC<PasswordSetProps> = ({
-  handleNext,
-  currentStep,
-  totalStep,
-  setPassword,
-}) => {
+const PasswordSet: React.FC = () => {
   const { t } = useTranslation();
   const classes = useStyles();
 
@@ -34,9 +38,9 @@ const PasswordSet: React.FC<PasswordSetProps> = ({
   const isError = useRef(true);
   const isSuccess = useRef(false);
   if (
-    values.password.length > 0 &&
-    values.confirmPassword.length > 0 &&
-    validateConfirmPassword(values.password, values.confirmPassword) === false
+    values.password.length &&
+    values.confirmPassword.length &&
+    !validateConfirmPassword(values.password, values.confirmPassword)
   ) {
     isError.current = false;
     isSuccess.current = true;
@@ -44,10 +48,90 @@ const PasswordSet: React.FC<PasswordSetProps> = ({
     isError.current = true;
     isSuccess.current = false;
   }
+  const username = getUsername();
 
-  const handleSubmit = () => {
-    handleNext();
+  const [loading, setIsLoading] = useState<boolean>(false);
+
+  // Mutation to create project for a user
+  const [CreateProject] = useMutation<Project>(CREATE_PROJECT, {
+    onCompleted: () => {
+      setIsLoading(false);
+      window.location.assign('/home');
+    },
+  });
+
+  // Mutation to create a user in litmusDB
+  const [CreateUser] = useMutation<CreateUserData>(CREATE_USER, {
+    onCompleted: () => {
+      CreateProject({
+        variables: {
+          projectName: `${username}'s project`,
+        },
+      });
+    },
+  });
+
+  // Query to fetch user details of user from litmusDB
+  const [getUserInfo] = useLazyQuery<CurrentUserDetails>(GET_USER_INFO, {
+    variables: { username },
+    // Adding the user to litmusDB if user does not exists
+    onError: (err) => {
+      if (err.message === 'mongo: no documents in result')
+        CreateUser({
+          variables: {
+            user: {
+              username,
+              email: getUserEmail(),
+              name: getUserFullName(),
+              role: getUserRole(),
+              userID: getUserId(),
+            },
+          },
+        });
+      else console.error(err.message);
+    },
+    // Creating project for the user
+    onCompleted: () => {
+      CreateProject({
+        variables: {
+          projectName: `${username}'s project`,
+        },
+      });
+    },
+  });
+
+  // Submit entered data to /update endpoint
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    setIsLoading(true);
+    fetch(`${config.auth.url}/update/details`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${getToken()}`,
+      },
+      body: JSON.stringify({
+        username,
+        email: getUserEmail(),
+        name: getUserFullName(),
+        password: values.password,
+      }),
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        if ('error' in data) {
+          isError.current = true;
+        } else {
+          getUserInfo();
+        }
+      })
+      .catch((err) => {
+        isError.current = true;
+        console.error(err);
+      });
   };
+
+  const loaderSize = 20;
 
   return (
     <div className={classes.rootDiv}>
@@ -55,7 +139,7 @@ const PasswordSet: React.FC<PasswordSetProps> = ({
         <img src="icons/LitmusLogoLight.svg" alt="litmus logo" />
         <Typography className={classes.HeaderText}>
           {' '}
-          {t('getStarted.welcome')} {getUsername()}!
+          {t('getStarted.welcome')} {username}!
         </Typography>
         <Typography className={classes.HeaderText}>
           {t('getStarted.password.info')}
@@ -64,7 +148,11 @@ const PasswordSet: React.FC<PasswordSetProps> = ({
           {t('getStarted.password.desc')}
         </Typography>
       </div>
-      <form id="login-form" className={classes.inputDiv}>
+      <form
+        id="login-form"
+        className={classes.inputDiv}
+        onSubmit={handleSubmit}
+      >
         <InputField
           data-cy="inputPassword"
           className={classes.inputValue}
@@ -78,7 +166,6 @@ const PasswordSet: React.FC<PasswordSetProps> = ({
               password: event.target.value,
               confirmPassword: values.confirmPassword,
             });
-            setPassword(event);
           }}
         />
         <InputField
@@ -102,20 +189,19 @@ const PasswordSet: React.FC<PasswordSetProps> = ({
           }
         />
         <div className={classes.buttonGroup}>
-          <div data-cy="nextButton">
+          <div data-cy="finishButton">
             <ButtonFilled
               className={classes.submitButton}
               type="submit"
               disabled={isError.current}
-              onClick={handleSubmit}
             >
-              {t('getStarted.button.continue')}
+              {loading ? (
+                <Loader size={loaderSize} />
+              ) : (
+                <>{t('getStarted.button.finish')}</>
+              )}
             </ButtonFilled>
           </div>
-          <Typography className={classes.step}>
-            {t('getStarted.button.step')} {currentStep}{' '}
-            {t('getStarted.button.of')} {totalStep}
-          </Typography>
         </div>
       </form>
     </div>
